+++ CloudShell.ResourceDefinitions.ReferenceProviders/LocalVolumeResourceTypeServiceCollectionExtensions.cs
@@ -24,6 +24,12 @@ public static class LocalVolumeResourceTypeServiceCollectionExtensions
             ServiceDescriptor.Singleton<IResourceChangeApplyProvider, LocalVolumeResourceTypeProvider>());
         services.TryAddEnumerable(
             ServiceDescriptor.Singleton<IResourceDefinitionApplyProvider, LocalVolumeResourceTypeProvider>());
+        services.TryAddEnumerable(
+            ServiceDescriptor.Singleton<IResourceOperationProvider, LocalVolumeProvisionOperationProvider>());
+        services.TryAddEnumerable(
+            ServiceDescriptor.Singleton<IResourceOperationProjector, LocalVolumeProvisionOperationProvider>());
+        services.TryAddEnumerable(
+            ServiceDescriptor.Singleton<IResourceProjectionProvider, LocalVolumeResourceProjectionProvider>());
 
         return services;
     }